fix(security): normalize object-shaped PR-review findings before rendering (#690)
pr_review_render.render_comment stringified whatever shape a blocking/
comment finding arrived in. The model sometimes returns a finding as a
plain string and sometimes as a structured object; stringifying an
object produced a raw Python dict repr in the posted PR comment,
truncated mid-value, so the finding text was lost rather than merely
ugly (PR #690, 2026-09-08T14:04:35Z).

Add _finding_text(item): a string passes through unchanged; a dict
with recognizable file/issue-like keys renders as "location - text"
prose; any other dict falls back to "key: value" prose. Never a raw
repr.

Also split pr_review_render.py (crossed the 300-line guideline) into
pr_review_render.py (comment assembly) and pr_review_sanitize.py
(sanitiser + model-facing metadata), mirroring the existing
pr_review_lib/pr_review_render split.

Run-ID: iterate-2026-09-09-pr-review-dict-finding-render
