+++ .shipwright/agent_docs/conventions.md
@@ -24,6 +24,7 @@ Python 3.11+ with uv as package manager. All scripts are invoked via uv run. Hoo
 ## Learnings
 
 > **One line per learning, no bold paragraphs** — always-loaded Layer-1 context. Format: `- (YYYY-MM-DD) <phase> — <one-line rule/gotcha>. → <run_id|ADR-NNN>`. **Budget ≤ 600 chars; if a learning won't fit one line it is a *decision* — write an ADR and leave a one-line pointer here.** Enforced repo-agnostically by the F11 verifier + `shared/scripts/tools/check_agent_doc_budget.py`; see `references/reflection.md`. Full rationale for compacted entries is in [`../planning/adr/_archive-agent-doc-updates.md`](../planning/adr/_archive-agent-doc-updates.md).
+- (2026-09-09) build — hand-authoring a control-character regex class (e.g. a zero-width/bidi Unicode escape range) directly in a `Write`/`Edit` content string is unsafe even outside a Bash heredoc: the model's own text generation can silently emit the real invisible character instead of the escape-sequence text, producing a byte-corrupt source file (`SyntaxError: source code string cannot contain null bytes`). Build such a literal from pure ASCII hex codepoints (`"\\x%02x" % cp`) in a `uv run python <script>` step instead of typing it. → iterate-2026-09-09-pr-review-dict-finding-render
 - (2026-09-08) review — a review-record disposition (e.g. `code-review.json` marking a finding "accepted-and-fixed") must be VERIFIED against the actual diff before being trusted, not just asserted from memory of intent; external code review caught one recorded as fixed that was never applied to the source. → iterate-2026-09-08-ci-provenance-attestation
 - (2026-09-08) design — a predicate named `verified`/`confirmed` that only checks structural agreement (not execution-tier correctness like tests/coverage) must say so explicitly in its own docstring and every call site's docs — an unqualified name reads as stronger proof than it provides. → iterate-2026-09-08-ci-provenance-attestation
 - (2026-09-07) hooks — a hook fix under `shared/scripts/hooks/` executes at runtime from the PLUGIN CACHE (`~/.claude/plugins/cache/shipwright/...`), never the worktree source; a live probe that edits/instruments only the worktree copy silently exercises stale cached code and can produce a false "feature X is absent" reading. Verify against a real hook payload by instrumenting the cache copy itself (with explicit user OK — it's outside the worktree), then restore it byte-for-byte. → iterate-2026-09-07-bloat-gate-subagent-marker-isolation
