--- .shipwright/agent_docs/iterates/iterate-2026-08-13-triage-detail-selfcap.json
@@ -1,11 +1,0 @@
-{
-  "run_id": "iterate-2026-08-13-triage-detail-selfcap",
-  "date": "2026-08-13T12:30:18.798939Z",
-  "event_at": "2026-08-13T09:58:22.933942+00:00",
-  "type": "change",
-  "complexity": "small",
-  "branch": "iterate/triage-detail-selfcap",
-  "spec": null,
-  "tests_passed": true,
-  "adr": "iterate-2026-08-13-triage-detail-selfcap"
-}
