+++ CHANGELOG-unreleased.d/Fixed/iterate-2026-09-09-pr-review-dict-finding-render_001.md
@@ -1,0 +1,1 @@
+Normalized PR review findings to prose before rendering, so an object-shaped finding from the model no longer appears as a truncated raw Python dict repr in the posted PR comment (#690).
